move connect button to its own library
